Atlantis.Framework.AuthAuthorize update to reference new auth service
git-tfs-id: [http://tfs2010:8080/tfs/gd]$/Delta/Services/AtlantisFramework/4.0;C255929

+++ Atlantis.Framework.AuthAuthorize/Atlantis.Framework.AuthAuthorize.Interface/AuthAuthorizeResponseData.cs
@@ -12,11 +12,11 @@ namespace Atlantis.Framework.AuthAuthorize.Interface
 
     public HashSet<int> ValidationCodes { get; private set; }
 
-    public int StatusCode { get; private set; }
+    public long StatusCode { get; private set; }
     
     public string StatusMessage { get; private set; }
 
-    public AuthAuthorizeResponseData(string responseXml, int statusCode, HashSet<int> validationCodes, string statusMessage)
+    public AuthAuthorizeResponseData(string responseXml, long statusCode, HashSet<int> validationCodes, string statusMessage)
     {
       ResponseXml = responseXml;
       StatusCode = statusCode;
